Combine Kokkos_Core_fwd.hpp and Kokkos_Types_fwd.hpp
